refactor/#6: Capsule 생성 Req, Res Dto명 변경
